emmiter logic into trait
